fix type errors as much as possible
