+++ .eslintrc.json
@@ -18,6 +18,7 @@
 		"@typescript-eslint/no-unsafe-argument": "off",
 		"@typescript-eslint/restrict-plus-operands": "off",
 		"prefer-const": "off",
-		"@typescript-eslint/no-non-null-assertion": "off"
+		"@typescript-eslint/no-non-null-assertion": "off",
+		"@typescript-eslint/no-empty-interface": "off"
 	}
 }
