+++ components/Tiptap.tsx
@@ -1,118 +1,11 @@
 import React from 'react'
-import { useEditor, EditorContent } from '@tiptap/react'
-import Document from '@tiptap/extension-document'
-import Mention from '@tiptap/extension-mention'
-import suggestion from './suggestion'
-import Link from '@tiptap/extension-link'
-import Tweet from '../extensions/Paragraph'
-import Highlight from '@tiptap/extension-highlight'
-import { History } from './History'
-import { KVStorage } from './KVStorage'
-import { CustomKeymap } from '../extensions/CustomKeymap'
-import { CustomClipboardTextSerializer } from './CustomClipboardTextSerializer'
-import StarterKit from '@tiptap/starter-kit'
-import createEmptyTweetEditorModel from '../utils/editorUtils/createEmptyTweetEditorModel'
-import UniqueID from '../extensions/UniqueIDTiptap'
+import { EditorContent } from '@tiptap/react'
+import useCustomEditor from '../hooks/useCustomEditor'
 
 const Tiptap = () => {
   // const limit = 280
-
-  const initialContent = {
-    type: 'doc',
-    content: [createEmptyTweetEditorModel()]
-  }
-
-  const editor = useEditor({
-    extensions: [
-      StarterKit.configure({
-        // nodes
-        text: true,
-        blockquote: false,
-        bulletList: false,
-        codeBlock: false,
-        document: false,
-        hardBreak: false,
-        heading: false,
-        horizontalRule: false,
-        listItem: false,
-        orderedList: false,
-        paragraph: {
-          HTMLAttributes: {
-            class: 'tweet-p',
-          }
-        },
-        // marks
-        bold: false,
-        code: false,
-        italic: false,
-        strike: false,
-        // extensions
-        dropcursor: {
-          class: 'drop-cursor'
-        },
-        gapcursor: false,
-        history: false
-      }),
-      Document.extend({ content: 'tweet*' }),
-      Tweet,
-      Highlight.configure({
-        multicolor: true,
-        HTMLAttributes: {
-          style: 'color: inherit'
-        }
-      }),
-      Mention.configure({
-        HTMLAttributes: {
-          class: 'mention'
-        },
-        suggestion
-      }),
-      Link.configure({
-        openOnClick: true
-      }),
-      CustomKeymap,
-      CustomClipboardTextSerializer,
-      History,
-      KVStorage,
-      UniqueID.configure({
-        types: ['tweet'],
-      }),
-    ],
-    content: initialContent,
-    autofocus: 'start'
-    // onUpdate({ editor }) {
-    //   editor.chain().focus(1)
-    // },
-    // onBeforeCreate: ({ editor }) => {
-    //   console.log('onBeforeCreate', editor)
-    // },
-    // onBeforeCreate({ editor }) {
-    //   console.log('onBeforeCreate', editor)
-    // },
-    // onCreate({ editor }) {
-    //   console.log('onCreate', editor)
-    // },
-    // onUpdate({ editor }) {
-    //   editor.commands.focus('start')
-    //   const value = editor.getText()
-    //   if (value.endsWith('\n\n')) {
-    //     console.log('Break', editor)
-    //   }
-    // }
-    // onSelectionUpdate({ editor }) {
-    //   console.log('onSelectionUpdate', editor)
-    // },
-
-    // onFocus({ editor, event }) {
-    //   console.log('onFocus', editor, event)
-    // },
-    // onBlur({ editor, event }) {
-    //   console.log('onBlur', editor, event)
-    // },
-    // onDestroy() {
-    //   console.log('onDestroy')
-    // }
-  })
+ 
+	const editor = useCustomEditor();
 
   if (!editor) {
     return null
