+++ extensions/UniqueIDTiptap.ts
@@ -118,7 +118,7 @@ const UniqueID: Extension<UniqueIDOptions, any> = Extension.create({
           const filterTransactions =
             this.options.filterTransaction &&
             transactions.some((tr) => {
-              var _a, _b;
+              let _a, _b;
               return !((_b = (_a = this.options).filterTransaction) === null ||
               _b === void 0
                 ? void 0
@@ -148,7 +148,7 @@ const UniqueID: Extension<UniqueIDOptions, any> = Extension.create({
               .filter((id) => id !== null);
             const duplicatedNewIds = findDuplicates(newIds);
             newNodes.forEach(({ node, pos }) => {
-              var _a;
+              let _a;
               // instead of checking `node.attrs[attributeName]` directly
               // we look at the current state of the node within `tr.doc`.
               // this helps to prevent adding new ids to the same node
@@ -183,7 +183,7 @@ const UniqueID: Extension<UniqueIDOptions, any> = Extension.create({
         // we register a global drag handler to track the current drag source element
         view(view) {
           const handleDragstart = (event) => {
-            var _a;
+            let _a;
             dragSourceElement = (
               (_a = view.dom.parentElement) === null || _a === void 0
                 ? void 0
@@ -206,7 +206,7 @@ const UniqueID: Extension<UniqueIDOptions, any> = Extension.create({
             // only create new ids for dropped content while holding `alt`
             // or content is dragged from another editor
             drop: (view, event) => {
-              var _a;
+              let _a;
               if (
                 dragSourceElement !== view.dom.parentElement ||
                 ((_a = event.dataTransfer) === null || _a === void 0
