+++ utils/editorUtils/textFromTweetNode.ts
@@ -1,7 +1,7 @@
 import { IRichTextTweet } from "../../types/IRichTextTweet";
 import { ISentence } from "../../types/ISentence";
 
-function textFromTweetNode(tweet: IRichTextTweet<"paragraph">) {
+function textFromTweetNode(tweet: IRichTextTweet) {
   const paragraphsText = (tweet.content ?? []).map((p) => {
     if (Array.isArray(p.content)) {
       return p.content.map((el) => el.text).join("");
